backend/requirements.txt
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
pillow==11.1.0
PyJWT==2.9.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycparser==2.22
django-environ==0.11.2
gunicorn==23.0.0
whitenoise==6.9.0

